+++ Server/R/CaseDataManager.R
@@ -399,6 +399,7 @@ CaseDataManager <- R6::R6Class( # nolint
               private$Catalogs$AdjustedData <- data
 
               private$InvalidateAfterStep('CASE_BASED_ADJUSTMENTS')
+              cat('\n')
               PrintAlert('Adjustment task finished')
               private$SendMessage(
                 'ADJUSTMENTS_RUN_FINISHED',
@@ -466,7 +467,7 @@ CaseDataManager <- R6::R6Class( # nolint
 
     # 7. Migration ---------------------------------------------------------------------------------
     RunMigration = function(
-      params = GetMigrantParams()
+      params = HivEstInfTime::GetMigrantParams()
     ) {
       if (!is.null(private$AppMgr) && !is.element(
         private$AppMgr$Steps['CASE_BASED_ORIGIN_GROUPING'],
@@ -494,7 +495,7 @@ CaseDataManager <- R6::R6Class( # nolint
               .Random.seed <- randomSeed # nolint
 
               input <- hivPlatform::PrepareMigrantData(data)
-              output <- hivPlatform::PredictInf(input, params)
+              output <- HivEstInfTime::PredictInf(input, params)
               data[output, ProbPre := i.ProbPre, on = .(UniqueId)]
 
               # Enrich data with extra dimensions from the input preparation step
@@ -825,8 +826,6 @@ CaseDataManager <- R6::R6Class( # nolint
         )
       }, ignoreInit = TRUE)
 
-
-      PrintAlert('CaseDataManager observers created')
       return(invisible(self))
     }
   ),