+++ Server/R/hivPlatform-package.R
@@ -9,12 +9,11 @@
 #' Improves accuracy of HIV estimates in EU/EEA countries.
 #'
 #' @author
-#' Author: Magdalena Rosinska \email{<mrosinska@@pzh.gov.pl>}\cr
-#' Author: Nikos Pantazis \email{<npantaz@@med.uoa.gr>}\cr
-#' Author: Ard van Sighem \email{<a.i.vansighem@amsterdamumc.nl>}\cr
-#' Creator: Daniel Lewandowski \email{<daniel@@nextpagesoft.net>}
+#' Author: Magdalena Rosinska \email{mrosinska@@pzh.gov.pl}\cr
+#' Author: Nikos Pantazis \email{npantaz@@med.uoa.gr}\cr
+#' Author: Ard van Sighem \email{a.i.vansighem@amsterdamumc.nl}\cr
+#' Creator: Daniel Lewandowski \email{daniel@@nextpagesoft.net}
 #'
-#' @importFrom Rcpp sourceCpp
 #' @importFrom grDevices dev.control dev.off hcl pdf recordPlot
 #' @importFrom grid grid.newpage grid.layout viewport pushViewport
 #' @importFrom stats setNames lowess quantile sd var acf glm poisson as.formula coef coefficients
@@ -25,5 +24,4 @@
 #' @import data.table
 #' @import ggplot2
 #'
-#' @useDynLib hivPlatform, .registration = TRUE
 NULL