+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.0.1
+
+###### _June 11 2023_
+
+- Move migrant functionality to package HivEstInfTime
+- Streamline logging of adjustments
+- Update Windows deployment package to R 4.3.0
+
+
 ## 3.0.0
 
 ###### _April 16 2023_
@@ -7,6 +16,7 @@
 - Move Github repo to https://github.com/nextpagesoft/hivPlatform
 - Update Windows deployment package to R 4.2.3
 - Release version 3.0.0
+
 
 ## 3.0.0